# Extend io::copy buffer reuse to BufReader too