@rust-log-analyzer 2023-06-05
The job **`mingw-check-tidy`** failed! Check out the build log: [(web)](https://github.com/rust-lang/rust/actions/runs/5182839587/jobs/9340055838) [(plain)](https://github.com/rust-lang/rust/commit/81c1d012c3a771154386c4efaa6ba062d65a91b3/checks/14024738569/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain
Prepare all required actions
Getting action download info
Download action repository 'actions/checkout@v3' (SHA:8e5e7e5ab8b370d6c329ec480221332ada57f0ab)
Download action repository 'rust-lang/simpleinfra@master' (SHA:aa3fbae38ebe51be080f7f515b11e5c847538bc2)
Complete job name: PR - mingw-check-tidy
git config --global core.autocrlf false
shell: /usr/bin/bash --noprofile --norc -e -o pipefail {0}
env:
  CI_JOB_NAME: mingw-check-tidy
---
Building wheels for collected packages: reuse
  Building wheel for reuse (pyproject.toml): started
  Building wheel for reuse (pyproject.toml): finished with status 'done'
  Created wheel for reuse: filename=reuse-1.1.0-cp310-cp310-manylinux_2_35_x86_64.whl size=180116 sha256=351235b2326fb4db7a18e257e13ce7896c5f77339521e2c2612e71e154800a19
  Stored in directory: /tmp/pip-ephem-wheel-cache-__8acec5/wheels/c2/3c/b9/1120c2ab4bd82694f7e6f0537dc5b9a085c13e2c69a8d0c76d
Installing collected packages: boolean-py, binaryornot, setuptools, reuse, python-debian, markupsafe, license-expression, jinja2, chardet
  Attempting uninstall: setuptools
    Found existing installation: setuptools 59.6.0
    Not uninstalling setuptools at /usr/lib/python3/dist-packages, outside environment /usr
---
Successfully tagged rust-ci:latest
Built container sha256:f1ea47ce5579ffb1112cdfba3bdddb64f3ada30d98ff1169c838eb30cb9b2dd2
Uploading finished image to https://ci-caches.rust-lang.org/docker/e51156f19850ce886cec818e46dc2a021e0aa7c270d15673e8fe74cd8522fc8ac3995109aebb688ee49ed586735ed4cf5f8c06d44c48298fa09c35ae4b082281

<botocore.awsrequest.AWSRequest object at 0x7f1912be0350>
gzip: stdout: Broken pipe
xargs: docker: terminated by signal 13
[CI_JOB_NAME=mingw-check-tidy]
[CI_JOB_NAME=mingw-check-tidy]
---
+use crate::io::BufReader;
 use crate::io::Cursor;
+use test::Bencher;
 
 #[test]
 fn test_copy_buf_reader_specialization() {
Running `"/checkout/obj/build/x86_64-unknown-linux-gnu/rustfmt/bin/rustfmt" "--config-path" "/checkout" "--edition" "2021" "--unstable-features" "--skip-children" "--check" "/checkout/library/std/src/io/prelude.rs" "/checkout/library/std/src/io/stdio.rs" "/checkout/library/std/src/io/impls.rs" "/checkout/library/std/src/io/readbuf/tests.rs" "/checkout/library/std/src/io/copy/tests.rs" "/checkout/library/std/src/io/readbuf.rs" "/checkout/library/std/src/io/error.rs" "/checkout/library/std/src/io/buffered/bufreader/buffer.rs"` failed.
If you're running `tidy`, try again with `--bless`. Or, if you just want to format code, run `./x.py fmt` instead.
Diff in /checkout/library/std/src/io/copy/tests.rs at line 10:
Diff in /checkout/library/std/src/io/copy/tests.rs at line 10:
-    let mut source = vec![0; 768*1024];
+    let mut source = vec![0; 768 * 1024];
     source[1] = 42;
-    let mut buffered = BufReader::with_capacity(256*1024, Cursor::new(&mut source));
+    let mut buffered = BufReader::with_capacity(256 * 1024, Cursor::new(&mut source));
 
     let mut sink = Vec::new();
     assert_eq!(crate::io::copy(&mut buffered, &mut sink).unwrap(), source.len() as u64);
Diff in /checkout/library/std/src/io/copy/tests.rs at line 21:
 fn bench_copy_buf_reader(b: &mut Bencher) {
     let mut file_in = File::open("/dev/zero").expect("opening /dev/zero failed");
     let dyn_in = &mut file_in as &mut dyn Read;
-    let mut reader = BufReader::with_capacity(256*1024, dyn_in.take(0));
-    let mut writer = OpenOptions::new().write(true).open("/dev/null").expect("opening /dev/null failed");
+    let mut reader = BufReader::with_capacity(256 * 1024, dyn_in.take(0));
+    let mut writer =
+        OpenOptions::new().write(true).open("/dev/null").expect("opening /dev/null failed");
-    const BYTES: u64 = 1024*1024;
+    const BYTES: u64 = 1024 * 1024;
 
 
     b.bytes = BYTES;
```

</details>